bl1: modify for memory test
As memory testing progressed, the message was changed
to a user-friendly message. The memory size was
automatically calculated without the need to enter
information.

+++ src/devices/memory/ddr3/ddr3_sdram.h
@@ -52,7 +52,7 @@ struct dram_device_info {
 	unsigned int column_size;
 	unsigned int row_size;
 	unsigned long bank_size;
-	unsigned short chip_size;
+	unsigned long chip_size;
 	unsigned long sdram_size;
 };
